gpt_start_langchain/requirements.txt
openai==1.109.1
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
scikit-learn==1.7.2
coloredlogs==15.0.1
fastapi
uvicorn
python-multipart
PyPDF2
streamlit
plotly
scikit-learn
requests